packages/ai-service/requirements.txt
flask
torch
torchvision
opencv-python
numpy
pillow
python-dotenv
gunicorn
watchdog
flask-cors
pytest
requests-mock
transformers

